make current color always zero
Current color will be always zero for evaluation function. Hence min max algorithm's max part will always looks for max score of white - black. On the other hand, min part looks for min of white - black. I made algorithm work on, black part looks for max value for black - white, and white part looks for max of white - black. And so for min. Hence in one part for black, it was looking for max and other part of black it was looking for min.